PRC-53: Group up add/create pages
Moved all the controllers and associated entities under "add" instead of "create" in preparation for them to be shared.
