+++ server/views/pages/contacts/add/checkAnswers.njk
@@ -9,7 +9,7 @@
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
-{% include '../common/partials/createContactBreadcrumbs.njk' %}
+{% include './partials/createContactBreadcrumbs.njk' %}
 {{ miniProfile(prisonerDetails) }}
 {% include '../../../partials/formErrorSummary.njk' %}
 <span class="govuk-caption-l">Manage Contacts</span>
