+++ server/views/pages/contacts/add/enterDob.njk
@@ -10,7 +10,7 @@
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
-{% include '../common/partials/createContactBreadcrumbs.njk' %}
+{% include './partials/createContactBreadcrumbs.njk' %}
 {{ miniProfile(prisonerDetails) }}
 {% include '../../../partials/formErrorSummary.njk' %}
 <span class="govuk-caption-l">Manage Contacts</span>
